+++ Bindings/OpenSimHeaders_simulation.h
@@ -146,9 +146,11 @@
 #include <OpenSim/Simulation/AssemblySolver.h>
 #include <OpenSim/Simulation/Reference.h>
 #include <OpenSim/Simulation/MarkersReference.h>
+#include <OpenSim/Simulation/BufferedMarkersReference.h>
 #include <OpenSim/Simulation/CoordinateReference.h>
 #include <OpenSim/Simulation/OrientationsReference.h>
 #include <OpenSim/Simulation/InverseKinematicsSolver.h>
+#include <OpenSim/Simulation/BufferedInverseKinematicsSolver.h>
 #include <OpenSim/Simulation/OpenSense/IMUPlacer.h>
 #include <OpenSim/Simulation/OpenSense/IMU.h>
 #include <OpenSim/Simulation/OpenSense/OpenSenseUtilities.h>